ejer2 sem3

+++ SEMANA1_C#/Program.cs
@@ -35,7 +35,7 @@ namespace SEMANA1_C_
       Console.Write("Ingrese numero y: ");
       int y = Convert.ToInt32( Console.ReadLine() );
 
-      double resu = x / y;
+      double resu = (double)x / (double)y;
 
       Console.WriteLine("Suma: "+(x+y));
       Console.WriteLine("Resta: " + (x - y));
